requirements.txt
appdirs==1.4.3
appnope==0.1.0
argcomplete==1.9.3
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
awscli==1.16.14
backcall==0.1.0
base58==1.0.0
bleach==2.1.4
boardgamegeek2==0.0.1
boto3==1.9.13
botocore==1.12.4
certifi==2018.8.24
cffi==1.11.5
cfn-flip==1.0.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
datasketch==1.2.7
decorator==4.3.0
docutils==0.14
durationpy==0.5
entrypoints==0.2.3
Flask==1.0.2
future==0.16.0
hjson==3.0.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.2
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
placebo==0.8.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
Pygments==2.2.0
python-dateutil==2.6.1
python-slugify==1.2.4
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.1
redis==2.10.6
requests==2.19.1
requests-cache==0.4.13
rsa==3.4.2
s3transfer==0.1.13
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
toml==0.9.6
tornado==5.1
tqdm==4.19.1
traitlets==4.3.2
troposphere==2.3.3
Twisted==17.5.0
Unidecode==1.0.22
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
wsgi-request-logger==0.4.6
zappa==0.46.2
zope.interface==4.5.0
